+++ examples/basic.rs
@@ -120,7 +120,7 @@ fn main() {
         screen_width: CONSOLE_WIDTH * 8,
         screen_height: CONSOLE_HEIGHT * 8,
         window_title: "my roguelike".to_owned(),
-        font_path: "terminal_8x8.png".to_owned(),
+        font_paths: vec!["terminal_8x8.png".to_owned()],
         vsync: true,
         fullscreen: false,
         show_cursor: true,